fix: many things because of pr review
