+++ tinlake-ui/.env.dev-example
@@ -13,5 +13,5 @@ INVESTOR_REQUEST_EMAIL=ask@centrifuge.io
 NEXT_PUBLIC_IPFS_GATEWAY=https://cloudflare-ipfs.com/ipfs/
 NEXT_PUBLIC_POOL_REGISTRY=0x8FE85CeAe6157C1dfcDD1c5ec99361c9722d97de
 NEXT_PUBLIC_ONBOARD_API_HOST=http://localhost:3100/
+NEXT_PUBLIC_MULTICALL_CONTRACT_ADDRESS=0x2cc8688c5f75e365aaeeb4ea8d6a480405a48d2a
 #NEXT_PUBLIC_POOLS_IPFS_HASH_OVERRIDE
-
