+++ tinlake-ui/.env.prod-example
@@ -13,4 +13,5 @@ INVESTOR_REQUEST_EMAIL=ask@centrifuge.io
 NEXT_PUBLIC_IPFS_GATEWAY=https://cloudflare-ipfs.com/ipfs/
 NEXT_PUBLIC_POOL_REGISTRY=0xddf1c516cf87126c6c610b52fd8d609e67fb6033
 NEXT_PUBLIC_ONBOARD_API_HOST=http://localhost:3100/
+NEXT_PUBLIC_MULTICALL_CONTRACT_ADDRESS=0xeefba1e63905ef1d7acba5a8513c70307c1ce441
 #NEXT_PUBLIC_POOLS_IPFS_HASH_OVERRIDE
