+++ tinlake-ui/ducks/loans.ts
@@ -14,10 +14,7 @@ const ZERO_ADDRESS = '0x0000000000000000000000000000000000000000'
 
 const multicallConfig = {
   rpcUrl: config.rpcUrl,
-  multicallAddress:
-    config.network === 'Mainnet'
-      ? '0xeefba1e63905ef1d7acba5a8513c70307c1ce441'
-      : '0x2cc8688c5f75e365aaeeb4ea8d6a480405a48d2a',
+  multicallAddress: config.multicallContractAddress,
   interval: 60000,
 }
